floats -> doubles

+++ src/main/java/ru/maklas/mengine/performace/ByIdCapture.java
@@ -3,9 +3,9 @@ package ru.maklas.mengine.performace;
 public class ByIdCapture extends Capture{
 
     long avgTimePerFrame;
-    float callsPerFrame;
+    double callsPerFrame;
 
-    public ByIdCapture(long averageNanoSeconds, long maxNanoSeconds, long minNanoSeconds, int framesChecked, long avgTimePerFrame, float callsPerFrame) {
+    public ByIdCapture(long averageNanoSeconds, long maxNanoSeconds, long minNanoSeconds, int framesChecked, long avgTimePerFrame, double callsPerFrame) {
         super(averageNanoSeconds, maxNanoSeconds, minNanoSeconds, framesChecked);
         this.avgTimePerFrame = avgTimePerFrame;
         this.callsPerFrame = callsPerFrame;
