+++ src/main/java/ru/maklas/mengine/performace/EventCapture.java
@@ -2,9 +2,9 @@ package ru.maklas.mengine.performace;
 
 public class EventCapture extends ClassCapture{
 
-    float callsPerFrame;
+    double callsPerFrame;
 
-    public EventCapture(long averageNanoSeconds, long maxNanoSeconds, long minNanoSeconds, int framesChecked, Class clazz, float callsPerFrame) {
+    public EventCapture(long averageNanoSeconds, long maxNanoSeconds, long minNanoSeconds, int framesChecked, Class clazz, double callsPerFrame) {
         super(averageNanoSeconds, maxNanoSeconds, minNanoSeconds, framesChecked, clazz);
         this.callsPerFrame = callsPerFrame;
     }
